Introduce the first working version
It's time I stopped working in my cave. :)

There's still a bit to do before this can actually be used by anyone. In
general there needs to be more testing and documentation. The API needs to
be finalised (especially w.r.t. the store).

This version only supports v3 of the protocol, but I indend to add v2
support. Support also needs to be added for group messaging and
simultaneous session initiations.