Refactor the interface to side-effect free
After attempting to use the interface for the first time, I found that it
was not very suitable. Implementing the large Store interface was also
painful.

The new interface is much more functional. All methods are side-effect free,
with the exception of calling the remaining store methods.

I've moved the trust check out of Axolotl as this should be done at a
higher level, in my opinion.

I've also removed the JSON serialisation of the session, as some clients may
want to serialise the session in other ways.